@@ -4,6 +4,8 @@ namespace App\FrontModule\Presenters;
 
 use Nette,
     Tracy\ILogger;
+use Nette\Application\BadRequestException;
+use Nette\Database\Context;
 
 /**
  * Error presenter.
@@ -14,7 +16,12 @@ class ErrorPresenter extends BasePresenter
     /** @var ILogger */
     private $logger;
 
-    public function __construct(ILogger $logger, Nette\Database\Context $database)
+    /**
+     * ErrorPresenter constructor.
+     * @param ILogger $logger
+     * @param Context $database
+     */
+    public function __construct(ILogger $logger, Context $database)
     {
         $this->logger = $logger;
         $this->database = $database;
@@ -26,10 +33,10 @@ class ErrorPresenter extends BasePresenter
      */
     public function renderDefault($exception)
     {
-        if ($exception instanceof Nette\Application\BadRequestException) {
+        if ($exception instanceof BadRequestException) {
             $code = $exception->getCode();
             // load template 403.latte or 404.latte or ... 4xx.latte
-            $this->setView(in_array($code, array(403, 404, 405, 410, 500)) ? $code : '4xx');
+            $this->setView(\in_array($code, [403, 404, 405, 410, 500], true) ? $code : '4xx');
             // log to access.log
             $this->logger->log("HTTP code $code: {$exception->getMessage()} in {$exception->getFile()}:{$exception->getLine()}", 'access');
         } else {
@@ -39,7 +46,7 @@ class ErrorPresenter extends BasePresenter
         }
 
         if ($this->isAjax()) { // AJAX request? Note this error in payload.
-            $this->payload->error = TRUE;
+            $this->payload->error = true;
             $this->terminate();
         }
     }
