@@ -1,6 +1,7 @@
 <?php
 namespace Caloriscz\Sign;
 
+use Nette\Application\AbortException;
 use Nette\Application\UI\Control;
 use Nette\Database\Context;
 
@@ -16,34 +17,34 @@ class VerifyAccountControl extends Control
     }
 
     /**
-     * @throws \Nette\Application\AbortException
+     * @throws AbortException
      */
-    public function handleCheck()
+    public function handleCheck(): void
     {
-        $userLoggedDb = $this->database->table('users')->where(array(
+        $userLoggedDb = $this->database->table('users')->where([
             'activation' => $this->getPresenter()->getParameter('code'),
             'username' => $this->getPresenter()->getParameter('user')
-        ));
+        ]);
 
         if ($userLoggedDb->count() === 0) {
             $this->getPresenter()->flashMessage('Aktivace není platná', 'error');
             $this->getPresenter()->redirect(':Front:Sign:verify');
         } else {
-            $this->database->table('users')->where(array(
+            $this->database->table('users')->where([
                 'activation' => $this->getPresenter()->getParameter('code'),
                 'username' => $this->getPresenter()->getParameter('user')
-            ))->update(array(
+            ])->update([
                 'state' => 1
-            ));
+            ]);
 
             $this->getPresenter()->flashMessage('Úspěšně ověřeno. Nyní se můžete přihlásit.', 'note');
             $this->getPresenter()->redirect(':Front:Sign:in');
         }
     }
 
-    public function render()
+    public function render(): void
     {
-        $template = $this->template;
+        $template = $this->getTemplate();
         $template->setFile(__DIR__ . '/VerifyAccountControl.latte');
 
         $template->render();
